Move `IdReadOnlyDescriptor` to another script file
