+++ tasktree/core/connection.py
@@ -2,23 +2,10 @@
 
 import re
 from typing import Iterable
-from tasktree.core.system import TaskSystem
+from tasktree.core.system import TaskSystem, IdReadOnlyDescriptor
+
 
 IdType = TaskSystem.IdType
-
-class IdReadOnlyDescriptor:
-    """Descriptor for id field, so you can not modify id of the established connection."""
-
-    def __get__(self, obj, cls):
-        """Returns the id."""
-        return obj._id_field
-
-    def __set__(self, obj, val):
-        """Sets the id. Works only once for each object."""
-        if hasattr(obj, '_id_set') and obj._id_set:
-            raise TypeError(f"'{type(obj).__name__}' object does not support id assignment")
-        obj._id_field = val
-        obj._id_set = True
 
 class Connection:
     """Connection between tasks."""
