+++ tasktree/core/system.py
@@ -4,6 +4,20 @@ import random
 import string
 
 
+class IdReadOnlyDescriptor:
+    """Descriptor for id field, so you can not modify id of the established connection."""
+
+    def __get__(self, obj, cls):
+        """Returns the id."""
+        return obj._id_field
+
+    def __set__(self, obj, val):
+        """Sets the id. Works only once for each object."""
+        if hasattr(obj, '_id_set') and obj._id_set:
+            raise TypeError(f"'{type(obj).__name__}' object does not support id assignment")
+        obj._id_field = val
+        obj._id_set = True
+
 class TaskSystem:
     """Control the system of tasks."""
